The email of the customer are now unique Some exceptions and checks were added to CustomerService

+++ src/main/java/it/unicam/cs/ids/loyaltyplatform/service/CustomerService.java
@@ -2,6 +2,7 @@ package it.unicam.cs.ids.loyaltyplatform.service;
 
 import it.unicam.cs.ids.loyaltyplatform.entity.users.Customer;
 import it.unicam.cs.ids.loyaltyplatform.repository.CustomerRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,19 +20,31 @@ public class CustomerService {
     }
 
     public void deleteCustomerById(Integer id) {
+        if(customerRepository.findAll().parallelStream().noneMatch(x -> x.getId().equals(id))) {
+            throw new EntityNotFoundException("The id(" + id + ") of the Customer to delete does not exist");
+        }
         this.customerRepository.deleteById(id);
     }
 
-    public void updateCustomer(Customer customer) {
-        this.customerRepository.saveAndFlush(customer);
+    public void updateCustomer(Integer id, Customer customer) {
+        Customer c = customerRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("This id(" + id + ") does not corresponds to any Customer"));
+        c.setName(customer.getName());
+        c.setSurname(customer.getSurname());
+        c.setGender(customer.getGender());
+        c.setEmail(customer.getEmail());
+        c.setPassword(customer.getPassword());
+        this.customerRepository.saveAndFlush(c);
     }
 
     public void addPointsToCustomerById(Integer id, Integer points) {
+        if(customerRepository.findAll().parallelStream().noneMatch(x -> x.getId().equals(id))) {
+            throw new EntityNotFoundException("The id(" + id + ") of the Customer does not exist");
+        }
         getCustomerById(id).sumPoints(points);
     }
 
     public Customer getCustomerById(Integer id) {
-        return this.customerRepository.findById(id).orElseThrow(NullPointerException::new);
+        return this.customerRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("The id(" + id + ") of the Customer to get does not exist"));
     }
 
     public List<Customer> getAllCustomers() {
